Fix name of DSL function reported in IncompatibleDSL
When the DSL uses a helper function such as `mediaTypeDefinition`. In
this case the caller depth is 3 instead of 3 when using direct `switch`
statements to lookup the current definition type. This fix moves the
helper functions to a different file so that IncompatibleDSL can infer
the correct stack depth.
